+++ orientation_test.go
@@ -2,6 +2,7 @@ package geom
 
 import (
 	"encoding/json"
+	"slices"
 	"testing"
 
 	"github.com/gravitton/assert"
@@ -101,12 +102,13 @@ func TestOrientation_Properties(t *testing.T) {
 		for _, orientation := range Orientations() {
 			for n := 3; n <= 8; n++ {
 				polygon := RegularPolygonWithOrientation(Pt(0.0, 0.0), SzU(10.0), n, orientation)
-				top := polygon.Vertices()[0]
+				vertices := slices.Collect(polygon.Vertices())
+				top := vertices[0]
 
 				if orientation == PointyTop {
 					AssertPoint(t, top, Pt(0.0, -10.0), orientation.String()+": ")
 				} else {
-					AssertVector(t, top.Midpoint(polygon.Vertices()[1]).Vector().Normalize(), Vec(0.0, -1.0), orientation.String()+": ")
+					AssertVector(t, top.Midpoint(vertices[1]).Vector().Normalize(), Vec(0.0, -1.0), orientation.String()+": ")
 				}
 			}
 		}